Silence some trivial compiler warning in the js2x code
We enabled additional compiler warnings in commit 26d8ba170355c9, but
so far we did not address the additional warnings in the js2x code
yet. This patch fixes at least some of the trivial warnings there.

+++ lib/libnvram/nvram.c
@@ -428,7 +428,7 @@ static partition_t create_nvram_partition(int type, const char *name, unsigned l
 
 	offset = calc_used_nvram_space();
 
-	if (NVRAM_LENGTH-(calc_used_nvram_space())<plen) {
+	if (NVRAM_LENGTH - calc_used_nvram_space() < (int)plen) {
 		DEBUG("Not enough free space.\n");
 		return ret;
 	}
